Added more documentation, reordered imports
