+++ backend/models.py
@@ -1,5 +1,7 @@
-from utilities import db
+from flask_sqlalchemy import SQLAlchemy
 from flask_login import UserMixin
+
+db = SQLAlchemy()
 
 class User(UserMixin, db.Model):
     # Keys
